Search Component
Converting courses search to search component and use it in the admin course page.
